% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corstar.R
\name{corstar}
\alias{corstar}
\title{A correlation table with stars}
\usage{
corstar(x, y = NULL, use = "pairwise", method = "pearson", round = 3,
  row.labels, col.labels, ...)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function allows you to create a correlation table with significance tables.
}
\examples{
corstar(cars)
}
\keyword{correlation}
\keyword{table}
