% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{create_dirs}
\alias{create_dirs}
\title{This function creates an any folder if one doesn't already exist.}
\usage{
create_dirs(folder)
}
\description{
This function creates an any folder if one doesn't already exist.
}
