% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{freqs}
\alias{freqs}
\title{Get Percentages and Frequencies}
\usage{
freqs(x, ...)
}
\arguments{
\item{x}{add in all grouping variables}
}
\value{
returns a dataframe with percentages and frequencies
}
\description{
This is a function that can be used in a pipe-workfolow
in order to get grouped frequencies and percentages
}
\examples{

iris \%>\% 
  freqs(Species)

}
