% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ggsave_it}
\alias{ggsave_it}
\title{This function saves a ggplot with the same name as the object.}
\usage{
ggsave_it(x, plot = ggplot2::last_plot(), ...)
}
\description{
This function saves a ggplot with the same name as the object.
}
