% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{images_dir}
\alias{images_dir}
\title{This function creates an images folder if one doesn't already exist.}
\usage{
images_dir()
}
\description{
This function creates an images folder if one doesn't already exist.
}
