% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logit2prob}
\alias{logit2prob}
\alias{logit2prob}
\title{Turns logits to probability}
\usage{
logit2prob(logit)

logit2prob(logit)
}
\description{
Turns logits to probability

Turns logits to probability
}
