% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruler.R
\name{ruler}
\alias{ruler}
\title{Measure the width of the console}
\usage{
ruler(width = getOption("width"))
}
\arguments{
\item{width}{Width of rule}
}
\description{
Measure the width of the console
}
\examples{
ruler()
}
