% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tbl_std}
\alias{tbl_std}
\title{Create table with standardized estimates}
\usage{
tbl_std(model, type = "text", ...)
}
\arguments{
\item{model}{put in the model (lm object) here.}

\item{...}{stargazer options.}
}
\value{
returns a stargazer table.
}
\description{
This function creates a stargazer table
with standardized estimates
}
\examples{
# #fake data
var1<-rnorm(100, mean=10, sd=5)
var2<-rnorm(100, mean=5, sd=2)
var3<-rnorm(100, mean=2, sd=3)
var4<-rnorm(100, mean=5, sd=1)
df<-data.frame(var1, var2, var3, var4)
#model with unstandardized betas
model1<-lm(var1~var2+var3+var4, data=df)

tbl_std(model1)
}
